+++ ts_jest/defaults.ts
@@ -1,2 +1,6 @@
-
 export const defaultWasmMemParams = {initial: 1, maximum: 1000, shared: false};
+export const defMsgPackOptions = {
+    useBigInt64: true,
+    ignoreUndefined: true,
+    initialBufferSize: 0,
+};