+++ ts_jest/errors.ts
@@ -1,3 +1,4 @@
+export const WASM_EXEC_ERROR = 'Smart contract execution error';
 export const WM_NOT_INST = 'Wasm machine not yet instantiated.';
 export const CONTRACT_NOT_REGISTERED = 'Contract is not registered';
 export const WRONG_ACCOUNT_CONTRACT = 'This account already has a different smart contract bound to it.';
@@ -6,3 +7,8 @@ export const ACCOUNT_ALREADY_BOUND = 'Account already bound.';
 export const MODULE_NOT_FOUND = 'Could not find required wasm module.';
 export const METHOD_NOT_FOUND = 'Method not found.';
 export const ALLOC_NOT_FOUND = 'Memory allocation method not found or not exported.';
+export const NOT_DEC = 'Invalid decimal string.';
+export const NOT_HEX = 'Invalid hex string.';
+export const NOT_B58 = 'Invalid base58 string.';
+export const NOT_B64 = 'Invalid base64 string.';
+export const NOT_B64URL = 'Invalid base64url string.';